added cast

+++ baselines/pois/pois.py
@@ -523,10 +523,10 @@ def learn(make_env, make_policy, *,
         tf.add_to_collection('asserts', tf.assert_positive(clustered_behavioral_pdf, name='clust_behavioral_pdf_positive'))
         # Compute the J
         ratio_clustered = clustered_target_pdf / clustered_behavioral_pdf
-        #ratio_reward = tf.cast(ratio_clustered, tf.float32) * reward_unique                             # ---- No cluster cardinality
-        ratio_reward = tf.cast(ratio_clustered, tf.float32) * reward_unique * trajectories_per_cluster   # ---- Cluster cardinality
-        #w_return_mean = tf.reduce_sum(ratio_reward) / tf.cast(max_index, tf.float32)                    # ---- No cluster cardinality
-        w_return_mean = tf.reduce_sum(ratio_reward) / tf.cast(n_episodes, tf.float32)                    # ---- Cluster cardinality
+        #ratio_reward = tf.cast(ratio_clustered, tf.float32) * reward_unique                                                  # ---- No cluster cardinality
+        ratio_reward = tf.cast(ratio_clustered, tf.float32) * reward_unique * tf.cast(trajectories_per_cluster, tf.float32)   # ---- Cluster cardinality
+        #w_return_mean = tf.reduce_sum(ratio_reward) / tf.cast(max_index, tf.float32)                                         # ---- No cluster cardinality
+        w_return_mean = tf.reduce_sum(ratio_reward) / tf.cast(n_episodes, tf.float32)                                         # ---- Cluster cardinality
         # Divergences
         ess_classic = tf.linalg.norm(ratio_reward, 1) ** 2 / tf.linalg.norm(ratio_reward, 2) ** 2
         sqrt_ess_classic = tf.linalg.norm(ratio_reward, 1) / tf.linalg.norm(ratio_reward, 2)
